+++ CHANGELOG.md
@@ -34,6 +34,8 @@ for now removed features.
 - **list-item**: update style to fix unnecessary padding-left on a single `wcs-list-item-property` 
 - **accordion**: fix the font-family not rendering properly
 - **doc**: improve `wcs-form-field`, `wcs-hint`, `wcs-label`, `wcs-error` and `wcs-editable-field` documentation
+- **textarea**: fix debounce behavior by using `wcsInput` event instead of `wcsChange`
+- **input**: fix debounce behavior by using `wcsInput` event instead of `wcsChange`
 - **counter**: improve keyboard focus style
 
 ### Security